pyproject.toml
[project]
name = "ai-agent-hackathon-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "dikky",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.0.0,<2.0.0)",
    "litellm (>=1.70.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
